Separated out TableCatalogItem from Csv, SDMX and AbsItt items
