+++ lib/Map/TableStructure.js
@@ -650,6 +650,25 @@ TableStructure.prototype.merge = function(table2) {
 };
 
 /**
+ * Sets the relevant active time column on the table structure, defaulting to the first time column present 
+ * unless the tableStyle has a 'timeColumn' property.
+ * @param {TableStyle} tableStyle The table style.
+ */
+TableStructure.prototype.setActiveTimeColumn = function(tableStyle) {
+    var tableStructure = this;
+    if (typeof tableStyle.timeColumn !== 'undefined') {
+        tableStructure.activeTimeColumn = tableStructure.getColumnWithNameIdOrIndex(tableStyle.timeColumn);
+        if (defined(tableStructure.activeTimeColumn) && tableStructure.activeTimeColumn.type !== VarType.TIME) {
+            tableStructure.activeTimeColumn = tableStructure.columnsByType[VarType.TIME][0];
+            throw new DeveloperError('TableStyle.timeColumn "' + tableStyle.timeColumn + '" is not a valid time column.');
+        }
+    } else {
+        tableStructure.activeTimeColumn = tableStructure.columnsByType[VarType.TIME][0];
+    }
+};
+
+
+/**
  * Destroy the object and release resources. Is this necessary?
  */
 TableStructure.prototype.destroy = function() {
