+++ lib/Models/AbsIttCatalogItem.js
@@ -19,14 +19,14 @@ var when = require('terriajs-cesium/Source/ThirdParty/when');
 var AbsConcept = require('../Map/AbsConcept');
 var arrayProduct = require('../Core/arrayProduct');
 var arraysAreEqual = require('../Core/arraysAreEqual');
-var CsvCatalogItem = require('./CsvCatalogItem');
 var inherit = require('../Core/inherit');
-var TerriaError = require('../Core/TerriaError');
 var overrideProperty = require('../Core/overrideProperty');
 var proxyCatalogItemUrl = require('./proxyCatalogItemUrl');
 var RegionMapping = require('../Models/RegionMapping');
+var TableCatalogItem = require('./TableCatalogItem');
 var TableColumn = require('../Map/TableColumn');
 var TableStructure = require('../Map/TableStructure');
+var TerriaError = require('../Core/TerriaError');
 var VarSubType = require('../Map/VarSubType');
 var VarType = require('../Map/VarType');
 
@@ -138,13 +138,13 @@ var allowedRegionCodes = ['AUS', 'STE', 'SA4', 'SA3', 'SA2', 'SA1', 'CED', 'LGA'
  *
  * @alias AbsIttCatalogItem
  * @constructor
- * @extends CsvCatalogItem
+ * @extends TableCatalogItem
  *
  * @param {Terria} terria The Terria instance.
  * @param {String} [url] The base URL from which to retrieve the data.
  */
 var AbsIttCatalogItem = function(terria, url) {
-    CsvCatalogItem.call(this, terria, url);
+    TableCatalogItem.call(this, terria, url);
 
     /**
      * Gets or sets the ID of the ABS dataset.  You can obtain a list of all datasets by querying
@@ -262,7 +262,7 @@ var AbsIttCatalogItem = function(terria, url) {
 
 };
 
-inherit(CsvCatalogItem, AbsIttCatalogItem);
+inherit(TableCatalogItem, AbsIttCatalogItem);
 
 defineProperties(AbsIttCatalogItem.prototype, {
     /**
@@ -318,14 +318,14 @@ defineProperties(AbsIttCatalogItem.prototype, {
  * share link.
  * @type {String[]}
  */
-AbsIttCatalogItem.defaultPropertiesForSharing = clone(CsvCatalogItem.defaultPropertiesForSharing);
+AbsIttCatalogItem.defaultPropertiesForSharing = clone(TableCatalogItem.defaultPropertiesForSharing);
 AbsIttCatalogItem.defaultPropertiesForSharing.push('filter');
 AbsIttCatalogItem.defaultPropertiesForSharing.push('regionConcept');
 AbsIttCatalogItem.defaultPropertiesForSharing.push('regionTypeConcept');
 AbsIttCatalogItem.defaultPropertiesForSharing.push('displayPercent');
 freezeObject(AbsIttCatalogItem.defaultPropertiesForSharing);
 
-AbsIttCatalogItem.defaultSerializers = clone(CsvCatalogItem.defaultSerializers);
+AbsIttCatalogItem.defaultSerializers = clone(TableCatalogItem.defaultSerializers);
 
 AbsIttCatalogItem.defaultSerializers.filter = function(item, json) {
     // Create the 'filter' that would start us off with the same active items as are currently shown.
@@ -358,6 +358,10 @@ AbsIttCatalogItem.prototype._load = function() {
     that._tableStructure = new TableStructure(that.name, {displayDuration: that._tableStyle.displayDuration});
     that._regionMapping = new RegionMapping(that, that._tableStructure, that._tableStyle);
     return rebuildData(that);
+};
+
+AbsIttCatalogItem.prototype.startPolling = function() {
+    // Polling is not available on ABS ITT items.
 };
 
 // TODO: Sometimes these errors do not actually reach the end user, because they are thrown inside a promise.
