+++ lib/Models/SdmxJsonCatalogItem.js
@@ -15,8 +15,6 @@ var loadJson = require('terriajs-cesium/Source/Core/loadJson');
 var when = require('terriajs-cesium/Source/ThirdParty/when');
 
 var arrayProduct = require('../Core/arrayProduct');
-// var arraysAreEqual = require('../Core/arraysAreEqual');
-var CsvCatalogItem = require('./CsvCatalogItem');
 var DisplayVariablesConcept = require('../Map/DisplayVariablesConcept');
 var inherit = require('../Core/inherit');
 var TerriaError = require('../Core/TerriaError');
@@ -24,6 +22,7 @@ var overrideProperty = require('../Core/overrideProperty');
 var proxyCatalogItemUrl = require('./proxyCatalogItemUrl');
 var RegionMapping = require('../Models/RegionMapping');
 var sdmxJsonLib = require('../ThirdParty/sdmxjsonlib');
+var TableCatalogItem = require('./TableCatalogItem');
 var TableColumn = require('../Map/TableColumn');
 var TableStructure = require('../Map/TableStructure');
 var VariableConcept = require('../Map/VariableConcept');
@@ -47,13 +46,13 @@ var VarType = require('../Map/VarType');
  *
  * @alias SdmxJsonCatalogItem
  * @constructor
- * @extends CsvCatalogItem
+ * @extends TableCatalogItem
  *
  * @param {Terria} terria The Terria instance.
  * @param {String} [url] The base URL from which to retrieve the data.
  */
 var SdmxJsonCatalogItem = function(terria, url) {
-    CsvCatalogItem.call(this, terria, url);
+    TableCatalogItem.call(this, terria, url);
 
     // The options that should be passed to TableColumn when creating a new column.
     this._columnOptions = undefined;
@@ -217,7 +216,7 @@ var SdmxJsonCatalogItem = function(terria, url) {
 
 };
 
-inherit(CsvCatalogItem, SdmxJsonCatalogItem);
+inherit(TableCatalogItem, SdmxJsonCatalogItem);
 
 defineProperties(SdmxJsonCatalogItem.prototype, {
     /**
@@ -273,11 +272,11 @@ defineProperties(SdmxJsonCatalogItem.prototype, {
  * share link.
  * @type {String[]}
  */
-SdmxJsonCatalogItem.defaultPropertiesForSharing = clone(CsvCatalogItem.defaultPropertiesForSharing);
+SdmxJsonCatalogItem.defaultPropertiesForSharing = clone(TableCatalogItem.defaultPropertiesForSharing);
 SdmxJsonCatalogItem.defaultPropertiesForSharing.push('selectedInitially');
 freezeObject(SdmxJsonCatalogItem.defaultPropertiesForSharing);
 
-SdmxJsonCatalogItem.defaultSerializers = clone(CsvCatalogItem.defaultSerializers);
+SdmxJsonCatalogItem.defaultSerializers = clone(TableCatalogItem.defaultSerializers);
 SdmxJsonCatalogItem.defaultSerializers.selectedInitially = function(item, json) {
     // Create the 'selectedInitially' that would start us off with the same active items as are currently shown.
     json.selectedInitially = {};
@@ -401,6 +400,10 @@ SdmxJsonCatalogItem.prototype._load = function() {
     } else {
         return loadAndBuildTable(this);
     }
+};
+
+SdmxJsonCatalogItem.prototype.startPolling = function() {
+    // Polling is not implemented on SDMX-JSON items (yet).
 };
 
 // An event listened triggered whenever the dataSource or regionMapping changes.
@@ -936,7 +939,7 @@ function loadAndBuildTable(item, url) {
         // Set the columns and the concepts before building the total column, because it uses them both.
         var totalColumns = buildTotalColumns(item, columnCombinations); // The region column can't be active, so ok not to pass it.
         var columns = item._tableStructure.columns.concat(totalColumns);
-        CsvCatalogItem.setActiveTimeColumn(item._tableStructure, item._tableStyle);
+        item._tableStructure.setActiveTimeColumn(item._tableStyle);
         // We want to update the columns, which can of course update the region column.
         // RegionMapping watches for changes in the active column and tries to redisplay it if so.
         // However it does not expect the region column to change - regionDetails does not auto-update its column.
